tests: convert test_on_battery to run from non-tests dir
